refactor: utils.EdgeQuery -> domain.GraphEdge
