+++ src/test/scala/domain/graph/GraphEdgeSpec.scala
@@ -1,26 +1,20 @@
-package utils.edge
+package domain.graph
 
+import domain.table.column._
 import domain.table.{TableList, TableName}
-import domain.table.column.{
-  ColumnLength,
-  ColumnList,
-  ColumnName,
-  ColumnTypeBoolean,
-  ColumnTypeDouble,
-  ColumnTypeInt
-}
 import org.apache.tinkerpop.gremlin.tinkergraph.structure.TinkerFactory
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
+import utils.edge.EdgeQuery
 
-class EdgeUtilitySpec extends AnyFunSpec with Matchers {
+class GraphEdgeSpec extends AnyFunSpec with Matchers {
   describe("toDdl") {
     it("get Database Column Information") {
       val graph = TinkerFactory.createModern().traversal()
       val edgeQuery = EdgeQuery(graph)
       val edge = edgeQuery.getEdgesList(0, 1).head
 
-      EdgeUtility.toDdl(edge) shouldBe TableList(
+      edge.toDdl shouldBe TableList(
         Map(
           TableName("edge") -> ColumnList(
             Map(
@@ -43,9 +37,7 @@ class EdgeUtilitySpec extends AnyFunSpec with Matchers {
       val edgeQuery = EdgeQuery(graph)
       val edge = edgeQuery.getEdgesList(0, 1).head
 
-      EdgeUtility.toDml(
-        edge
-      ) shouldBe "INSERT INTO edge (in_v_id, out_v_id, property_weight, label_knows) VALUES (2, 1, 0.5, true);"
+      edge.toDml shouldBe "INSERT INTO edge (in_v_id, out_v_id, property_weight, label_knows) VALUES (2, 1, 0.5, true);"
     }
   }
 }
